Testing script var redefine + cleanup
